[pjrt] Route non-contiguous host buffers through the owned-tensor path (#5059)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/8622

### Problem description
Described in issue.

### What's changed
Detect whether the incoming host buffer is C-contiguous from its
`byte_strides` (`isDenseRowMajor`) and, when it is not (and it's not a
complex dtype), force the owned path

### Checklist
- [x] tt-mlir uplift where [createOwnedHostTensor honors the stride
argument](https://github.com/tenstorrent/tt-mlir/pull/8697)

+++ pjrt_implementation/src/api/loaded_executable_instance.cc
@@ -263,8 +263,7 @@ tt_pjrt_status LoadedExecutableInstance::createDefaultOutputBuffers(
     // product of all following dimension sizes.
     std::vector<std::int64_t> strides(output_shape->size());
     std::exclusive_scan(output_shape->rbegin(), output_shape->rend(),
-                        strides.rbegin(), std::uint32_t(1),
-                        std::multiplies<>());
+                        strides.rbegin(), std::int64_t(1), std::multiplies<>());
 
     // Only create host tensor when not in compile-only mode
     std::optional<tt::runtime::Tensor> host_tensor;
